gpio: ts7800v2: direction_output() set WiFi IO
Previously, the enable and reset pins for WiFi would be opened,
set to output, and then their states set. As the WiFi driver
evolved, it relies on the kernel paradigm of setting the output
value when setting the direction. This GPIO driver did not
implement this process for the WiFi IO