wilc: Re-locate IRQ setup after firmware start
Workaround for a hardware shortcoming on the TS-7800-V2 that
previously was not a problem in the older 14.x WILC driver.

WILC IRQ# is connected to CPU pin that only supports an internal
PD with no way to disable and no external PU resistor. This causes
IRQ# to float around 1.4 V and generates an IRQ storm once enabled

The fix holds off on grabbing the IRQ# GPIO until the WILC FW is
loaded and started. Once started, WILC FW puts IRQ# in push-pull
and the IRQ storm is avoided.

The correct fix would probably be to split this apart. At module
load time, request the GPIOs needed for IO and IRQ, and then in
wlan_initialize() load and start FW. Finally, request/enable IRQ.