firmware: Removed WILC firmware from tree
This broke kernel paradigms and ended up causing a later build
failure if 'make clean' was run. The kernel is not happy with
binary files in the tree, and clean removes them.

Correct firmware will be present in future images that use this
kernel. The firmware can also be added to existing images as
needed.